+++ CHANGELOG.md
@@ -20,6 +20,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - SCS: handle infinite values in linear inequalities
+- qtqp: handle infinite values in linear inequalities
 
 ## [4.13.0] - 2026-07-19